refactor(providers): share tool-call sanitizer helpers
- Move malformed tool-call formatting out of the OpenAI module
- Reuse the shared helper from OpenAI and Anthropic projections
